% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beacon.R
\name{beacon}
\alias{beacon}
\title{Get beacon result for a given timestamp}
\usage{
beacon(ds)
}
\arguments{
\item{ds}{timestamp to get}
}
\value{
beacon information
}
\description{
Get beacon result for a given timestamp
}
\examples{
beacon(1500679537)
}
