% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beacon.R
\name{beacon_previous}
\alias{beacon_previous}
\title{Get previous beacon record}
\usage{
beacon_previous(ds)
}
\arguments{
\item{ds}{timestamp to get}
}
\value{
previous beacon record
}
\description{
Get previous beacon record
}
