% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beacon.R
\name{beacon_start_chain}
\alias{beacon_start_chain}
\title{Get start of chain record}
\usage{
beacon_start_chain(ds)
}
\arguments{
\item{ds}{timestamp to get}
}
\value{
start of chain record
}
\description{
Get start of chain record
}
