% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_beacon_url}
\alias{process_beacon_url}
\title{Parse beacon url}
\usage{
process_beacon_url(beacon_url)
}
\arguments{
\item{beacon_url}{url to parse}
}
\value{
parsed result from a given URL on beacon website
}
\description{
Parse beacon url
}
