% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_timestamp}
\alias{process_timestamp}
\title{Process timestamp}
\usage{
process_timestamp(ds)
}
\arguments{
\item{ds}{timestamp or beacon result}
}
\value{
timestamp
}
\description{
Process timestamp
}
\examples{
process_timestamp("abc")
}
